Changed file name from split words to cap words
