+++ frontend/src/functions/cap_words.js
